Move factory to SEO module
