+++ projects/storefrontlib/src/lib/seo/index.ts
@@ -1,11 +1,2 @@
-import { Injector } from '@angular/core';
-import { SeoTitleService } from './seo-title.service';
-
 export * from './seo-title.service';
-
-export function initSeoService(injector: Injector) {
-  return function() {
-    const service = injector.get(SeoTitleService);
-    service.initPageTitle();
-  };
-}
+export * from './seo.module';
